smol refactor and rename
