+++ src/sentry/workflow_engine/processors/data_condition_group.py
@@ -114,7 +114,9 @@ def evaluate_data_conditions(
         return ProcessedDataConditionGroup(logic_result=True, condition_results=[])
 
     for condition, value in conditions_to_evaluate:
-        evaluation_result = condition.evaluate_value(value)
+        evaluation_result = condition.evaluate_value(
+            value
+        )  # TODO maybe log here when we don't get one back? and which type of dc it is
         is_condition_triggered = evaluation_result is not None
 
         if is_condition_triggered:
