+++ tests/sentry/incidents/endpoints/validators/test_validators.py
@@ -6,7 +6,7 @@ from sentry import audit_log
 from sentry.incidents.grouptype import MetricIssue
 from sentry.incidents.metric_alert_detector import (
     MetricAlertComparisonConditionValidator,
-    MetricAlertsDetectorValidator,
+    MetricIssueDetectorValidator,
 )
 from sentry.incidents.models.alert_rule import AlertRuleDetectionType
 from sentry.incidents.utils.constants import INCIDENTS_SNUBA_SUBSCRIPTION_TYPE
@@ -138,7 +138,7 @@ class TestMetricAlertsDetectorValidator(BaseValidatorTest):
 
     @mock.patch("sentry.workflow_engine.endpoints.validators.base.detector.create_audit_entry")
     def test_create_with_valid_data(self, mock_audit):
-        validator = MetricAlertsDetectorValidator(
+        validator = MetricIssueDetectorValidator(
             data=self.valid_data,
             context=self.context,
         )
@@ -198,8 +198,8 @@ class TestMetricAlertsDetectorValidator(BaseValidatorTest):
         )
 
     def test_invalid_detector_type(self):
-        data = {**self.valid_data, "type": "invalid_type"}
-        validator = MetricAlertsDetectorValidator(data=data, context=self.context)
+        data = {**self.valid_data, "detectorType": "invalid_type"}
+        validator = MetricIssueDetectorValidator(data=data, context=self.context)
         assert not validator.is_valid()
         assert validator.errors.get("type") == [
             ErrorDetail(
@@ -236,7 +236,7 @@ class TestMetricAlertsDetectorValidator(BaseValidatorTest):
                 ],
             },
         }
-        validator = MetricAlertsDetectorValidator(data=data, context=self.context)
+        validator = MetricIssueDetectorValidator(data=data, context=self.context)
         assert not validator.is_valid()
         assert validator.errors.get("nonFieldErrors") == [
             ErrorDetail(string="Too many conditions", code="invalid")
